Further revise username and project completion
This reverts most of the code to the "simpler" calling convention. It
also drops the initial rank-based ordering (which didn't appear to
influence the result ordering).

The other significant change is that we now quote the query string to
avoid stemming, which keeps the results more focused on the original
search term.

The change also includes a bug fix for the previous change. We once
again avoid sending the `contraints[query]` parameter when we don't have
a query term (i.e. it's empty).